requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.13.1
dj_rest_auth==2.2.4
django-allauth==0.54.0
djangorestframework-simplejwt
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
django-cors-headers==3.12.0
numpy==1.22.4
scipy==1.8.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3